chore: move components to v1
