+++ src/app/(frontend)/(auth)/confirm-signup/page.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useReducer } from "react";
 
-import BackLink from "@/components/BackLink";
-import Spinner from "@/components/Spinner";
+import BackLink from "@/components/v1/BackLink";
+import Spinner from "@/components/v1/Spinner";
 import { useI18n } from '@/hooks/useI18n';
 import { supabase } from "@/libs/supabase/client";
 import AuthService from "@/services/auth";
