+++ src/app/(frontend)/(payment)/payments/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useSearchParams } from "next/navigation";
 
-import BackLink from "@/components/BackLink";
+import BackLink from "@/components/v1/BackLink";
 import { useI18n } from "@/hooks/useI18n";
 
 const PaymentStatus = () => {
