+++ src/components/v1/FooterAuthScreen.tsx
