+++ src/components/v1/PasswordStrength.tsx
