+++ src/components/v1/Pricing/index.tsx
