+++ src/hooks/useCheckout.ts
@@ -1,4 +1,4 @@
-import { Plan } from "@/components/Pricing/PlanCard";
+import { Plan } from "@/components/v1/Pricing/PlanCard";
 import { FIXED_CURRENCY } from "@/constants/fixed-currency";
 import { HAS_FREE_TRIAL } from "@/constants/has-free-trial";
 import { useToast } from "@/hooks/useToast";
